Create com.jamfsoftware.priority.plist
This is pretty much a copy of the LaunchDaemon that JAMF installs when you a) enroll or b) modify check-in frequency.
Allows for escalation of check-in frequency using a custom trigger.  
Multiple policies can be assigned a common trigger.
Consider versions of this with 2,5,10,etc check-in frequencies.  
Could also be useful for terminating "stuck" recurring check-ins.